+++ app/definitions/field-validators/medical-centre.js
@@ -14,19 +14,21 @@ module.exports = {
       errorMsg: 'medical-centre:name.errors.required',
     }),
   ]),
+  address: SimpleField([
+    validatePostalAddress.bind({
+      errorMsgAddress1: 'medical-centre:address.address1.errors.required',
+      errorMsgAddress1Format: 'medical-centre:address.address1.errors.badFormat',
+      errorMsgAddress2Format: 'medical-centre:address.address2.errors.badFormat',
+      errorMsgPostcodeFormat: 'medical-centre:postcode.errors.badFormat',
+      altRegex: /^[^0-9]*$/,
+    }),
+  ]),
   phoneNumber: SimpleField([
     rules.required.bind({
       errorMsg: 'medical-centre:phoneNumber.errors.required',
     }),
     validatePhoneNumber.bind({
       errorMsg: 'medical-centre:phoneNumber.errors.badFormat',
-    }),
-  ]),
-  address: SimpleField([
-    validatePostalAddress.bind({
-      errorMsgAddress1: 'medical-centre:address.address1.errors.required',
-      errorMsgPostcodeFormat: 'medical-centre:postcode.errors.badFormat',
-      altRegex: /^[^0-9]*$/,
     }),
   ]),
   doctor: SimpleField([