+++ test/unit/routes/complete.spec.js
@@ -96,104 +96,6 @@ describe('Complete route', () => {
     },
   };
 
-  const covidReq = {
-    language: 'en',
-    csrfToken: () => '',
-    i18nTranslator: ({
-      t: () => this,
-      getLanguage: () => 'en',
-    }),
-    session: {
-      cyaVisited: true,
-      save: (cb) => {
-        cb(new Error('save error'));
-      },
-      employmentGather: [
-        {
-          employerName: 'Smiths',
-          frequency: 'weekly',
-          workTypes: ['employee'],
-        },
-        {
-          employerName: 'Browns',
-          frequency: 'monthly',
-          workTypes: ['subContractor'],
-        },
-      ],
-      pensionGather: {},
-      insuranceGather: {},
-      journeyData: {
-        getDataForPage: () => ({
-          conditions: [],
-          nino: 'test',
-          sortCode: {
-            sortCode: '010101',
-          },
-          claimStartDate: {
-            dd: '1',
-            mm: '1',
-            yyyy: '1111',
-          },
-          dateOfBirth: {
-            dd: '1',
-            mm: '1',
-            yyyy: '1111',
-          },
-          address: {
-            address1: 'test',
-            address2: 'test',
-            address3: 'test',
-            postcode: 'test',
-          },
-          employed: {
-            employed: 'yes',
-          },
-          'statutory-sick-pay': {
-            ssp: 'yes',
-          },
-          'statutory-sick-pay-recent': {
-            sspRecent: 'yes',
-          },
-        }),
-      },
-    },
-    journeyData: {
-      getDataForPage: () => ({
-        conditions: [],
-        nino: 'test',
-        sortCode: {
-          sortCode: '010101',
-        },
-        claimStartDate: {
-          dd: '1',
-          mm: '1',
-          yyyy: '1111',
-        },
-        coronavirusReasonForClaim: 'yes',
-        dateOfBirth: {
-          dd: '1',
-          mm: '1',
-          yyyy: '1111',
-        },
-        address: {
-          address1: 'test',
-          address2: 'test',
-          address3: 'test',
-          postcode: 'test',
-        },
-        employed: {
-          employed: 'yes',
-        },
-        'statutory-sick-pay': {
-          ssp: 'yes',
-        },
-        'statutory-sick-pay-recent': {
-          sspRecent: 'yes',
-        },
-      }),
-    },
-  };
-
   const res = {
     status: () => ({
       render: () => {
@@ -215,24 +117,6 @@ describe('Complete route', () => {
     router.get = (path, callback) => {
       assert.equal(path, '/complete');
       callback(req, res);
-    };
-    completeRoute(casaApp, mountUrl, router);
-  });
-
-  it('should set up a GET route, and render the complete coronavirus page when covid claim', (done) => {
-    casaApp.endSession = () => Promise.resolve();
-    res.redirect = () => done();
-    res.render = (template) => {
-      try {
-        assert.equal(template, 'pages/complete-coronavirus.njk');
-        done();
-      } catch (e) {
-        done(e);
-      }
-    };
-    router.get = (path, callback) => {
-      assert.equal(path, '/complete');
-      callback(covidReq, res);
     };
     completeRoute(casaApp, mountUrl, router);
   });
@@ -412,24 +296,6 @@ describe('Complete route', () => {
       router.get = (path, callback) => {
         assert.equal(path, '/complete');
         callback(testReq, res);
-      };
-      completeRoute(casaApp, mountUrl, router);
-    });
-
-    it('should set displaySSP1Content to false when \'yes\' is selected on /coronavirus page', (done) => {
-      casaApp.endSession = () => Promise.resolve();
-      res.redirect = () => done();
-      res.render = (template, { displaySSP1Content }) => {
-        try {
-          assert.isFalse(displaySSP1Content);
-          done();
-        } catch (e) {
-          done(e);
-        }
-      };
-      router.get = (path, callback) => {
-        assert.equal(path, '/complete');
-        callback(covidReq, res);
       };
       completeRoute(casaApp, mountUrl, router);
     });